+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Improvements
 - Add more opacity support for image overlays ([#761](../../pull/761))
 - Make annotation schema more uniform ([#763](../../pull/763))
+- Improve TileSource class repr ([#765](../../pull/765))
 
 ## Version 1.10.0